requirements.txt

beautifulsoup4==4.12.2
Flask==3.1.0
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.31.0
eventlet>=0.30.2
flask_cors==5.0.1